requirements.txt
# Core Framework
Flask==3.0.0
Flask-Caching==2.1.0
gunicorn==21.2.0

# HTTP Requests
requests>=2.31.0

# Firebase & Database
firebase-admin==6.2.0

# Google Authentication
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Environment & Configuration
python-dotenv==1.0.0

# Security Dependencies
cryptography>=41.0.0
pyJWT>=2.8.0
Werkzeug>=3.0.0

# Input Validation & Sanitization
bleach>=6.0.0
html5lib>=1.1
Flask-Caching>=2.1.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov>=4.1.0

# Code Quality
flake8>=6.0.0
black>=23.0.0

# Monitoring & Logging
structlog>=23.1.0
